vos: Check end-of-dump magic before deleting volume
When restoring a volume with a regular file, vos restore verifies the dump
file is terminated with the end-of-dump magic number before sending the
dump to the volume server.  Unfortunately, this check is done after the
target volume is deleted (which happens when doing a full restore of a
volume), in which case the target volume is unnecessarily removed.

Currently, UV_RestoreVolume2() invokes the WriteData() function as
callback to transmit the dump stream to the volume server. WriteData()
opens the usd file handle, performs the end-of-dump magic check (if the
file is seekable), sends the data using the SendFile() helper function,
then closes the dump file handle.

Instead, open and check the dump file magic before calling
UV_RestoreVolume2().  Pass the usd file handle and a pointer to the
SendFile() function to UV_RestoreVolume2() to send the dump data to the
volume server.

Move the code to perform the dump file magic check to the new
CheckDumpFile() helper function and remove the now unneeded WriteData()
function.

Change-Id: Ia04f378540b40c2e6360aa120592fbd793c33cae
Reviewed-on: https://gerrit.openafs.org/14710
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>