backup: Remove obsolete test programs
The code for the butc_test and ttest programs look to be developer
scaffolding for an earlier version of the backup server.  Those programs
use a version of the TC_PerformDump RPC which pre-dates OpenAFS, so the
butc_test and ttest have never been buildable in OpenAFS.

Remove this obsolete scaffolding code since it is not maintained and
trigger errors when scanned with static analyzers.

Change-Id: Ia3d0c8d3758e1ed41d285beafa64c326dc11658d
Reviewed-on: https://gerrit.openafs.org/15097
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>