backup: Make backup tests build again
The test programs in bucoord, butc, and butm have bit rotted and no
longer build.  Update the code to fix error and warnings, and remove
dead code.  Update the relevant makefiles so these test programs build
by default to avoid bitrot in the future.

With this change the test programs will build and run, but a tape
drive device is required to perform testing.

Change-Id: I978b98fe63fe6051409fc160e8c7d43050bcf707
Reviewed-on: https://gerrit.openafs.org/15098
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>