rx: Only use printf in the AIX kernel
An earlier version of the OpenAFS code used an unusual preprocessor
macro for the function osi_Msg and osi_VMsg.  This causes problems
on newer Linux kernels.  The macro was replaced with a function
leveraging vprintf, etc.  These functions are not available in the
AIX kernel, where our only option is printf.  However, AIX does
provide these functions in libc.  This change ensures we only use
printf when building the AIX kernel module.

Change-Id: Ib46c4712516846a489cc66fda0833907612b3316
Reviewed-on: https://gerrit.openafs.org/15118
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>