afs: Use strlcat instead of snprintf to enable AIX build
The original code used strcpy(), etc.  Some compilers have begun
to raise warnings that the use of these functions is unsafe, and
in the case of newer Linux kernels in particular, have caused the
build to fail outright.  To adapt to these compilers/kernels, the
code was changed to leverage strlcpy(), snprintf(), etc. as they
are safer.  AIX does not provide these functions within the
kernel.  For some of the functions such as strlcpy(), etc. we
can use replacement functions from libroken within the AIX kernel
extension.  However, libroken does not offer a suitable
replacement for snprintf(), and instead the same purpose can be
accomplished by using strlcpy() and multiple calls to strlcat().

Change-Id: Iae4b9d945b12dea88727a3453d06b8ff1eb285ae
Reviewed-on: https://gerrit.openafs.org/15122
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>