dumptool: Remove unused 'sflag' variable
clang-16 is flagging unused-but-set variables which result in build
errors when --enable-warning is turned on.

   ./dumptool.c:1285:54: error: variable 'sflag' set but not used
      [-Werror,-Wunused-but-set-variable]
       int errflg = 0, lflag = 0, iflag = 0, Fflag = 0, sflag = 0, Rflag = 0;

Within dumptool the '-s' option of the 'ls' subcommand sets the 'sflag'
variable, but the variable itself is never used anywhere else.

Remove the variable 'sflag', and ignore the '-s' option.

The sflag variable was unused in the commit that introduced the code:
"test-suite-pull-tools-directly-in-20020114" (fcb32c07ed6d)

Change-Id: I0769b9d3f76b4a802109628496bf332618162f89
Reviewed-on: https://gerrit.openafs.org/15181
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>