rx: Remove event_handler rx_pthread_* counters
The event_handler() function contains several counters that are
set and incremented but are not actually used.  Clang-16 is flagging
these with a 'unused-but-set-variable' warning resulting in build errors
when --enable-warnings is used.

rx_pthread.c:136:19: error: variable 'rx_pthread_n_event_expired' set
       but not used [-Werror,-Wunused-but-set-variable]
      unsigned long rx_pthread_n_event_expired = 0;

Several 'if' statements are used to conditionally set these variables,
but serve no other purpose. These 'if' statements can also be
eliminated.

Remove the unused rx_pthread_* counter variables as well as the
associated 'if' statements.  In addition the only purpose of the
'error' variable was to conditionally increment one of the removed
variables, remove the 'error' variable.

These variables have been 'unused but set' since the original commit
from IBM.

Change-Id: Ibad026cef2030c89a9421b1eea64049ef38fe37e
Reviewed-on: https://gerrit.openafs.org/15175
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>