Linux: Fix functions without prototypes
The Linux kernel build has -Wstrict-prototypes enabled and flags
functions without a proper prototype.  With clang-16 these warnings are
reported as errors when the kernel is configured with CONFIG_WERROR or
when openafs is configured with --enable-checking.

    src/libafs/MODLOAD-6.1.0-rc3-SP/osi_alloc.c:339:32: error: a
       function declaration without a prototype is deprecated in all
       versions of C [-Werror,-Wstrict-prototypes]
   osi_linux_verify_alloced_memory()
                                  ^
                                   void

Update function prototypes to comply with -Wstrict-prototypes.

There are no functional changes with this commit.

Change-Id: I31aeb6e40bd3c6d9bf519d80bbed9b031b99df84
Reviewed-on: https://gerrit.openafs.org/15194
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>