clang-16: Fix unused-but-set 'for' control vars
clang-16 is flagging unused-but-set variables which result in build
errors when --enable-warning is turned on.  Several of these variables
are used within 'for(..)' loops but are not actually referenced anywhere
else.

  fcrypt.c:181:16: error: variable 'i' set but not used
     [-Werror,-Wunused-but-set-variable]
    afs_uint32 i, j;
               ^
Refactor the for statements to eliminate the unused variables.

Note, this commit is a little more complex than simply removing unused
variables.

Change-Id: I4d2de60dffc3e717d784e820f2d91470c5e7926b
Reviewed-on: https://gerrit.openafs.org/15176
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>