clang-16: Fix unused-but-set 'for' control vars
clang-16 is flagging unused-but-set variables which result in build
errors when --enable-warning is turned on.  Several of these variables
are used within 'for(..)' loops but are not actually referenced anywhere
else.

  fcrypt.c:181:16: error: variable 'i' set but not used
     [-Werror,-Wunused-but-set-variable]
    afs_uint32 i, j;
               ^
Refactor the for statements to eliminate the unused variables.

Note, this commit is a little more complex than simply removing unused
variables.

Reviewed-on: https://gerrit.openafs.org/15176
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
(cherry picked from commit 05eb420829f9bca155c1d5af9fe96844db974770)

Change-Id: I80506cb8ae1c4bef67225eb11e3f640e8de52d26
Reviewed-on: https://gerrit.openafs.org/15271
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Harish Sharma <Harish.Sharma1@ibm.com>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>