Fix various warnings on RHEL5
Building on RHEL5 with gcc 4.1 yields a few warnings that break the
build with --enable-checking.

A couple of headers lack a newline at the end of the file, such as
audit.h:

      CC  .../src/audit/audit.lo
    In file included from .../src/audit/audit.c:30:
    .../src/audit/audit.h:337:31: error: no newline at end of file

And UKERNEL's afsincludes.h:

      CC  .../src/libuafs/afs_analyze.lo
    In file included from .../src/afs/afsincludes.h:16,
                     from .../git-foo/src/afs/afs_analyze.c:34:
    .../src/afs/UKERNEL/afsincludes.h:35:47: error: no newline at end of file

To fix these, just add a trailing newline.

osi_vfsops.c has an 'out:' label that is only used under certain
 #ifdefs, but we always define it, leading to a warning when it's
defined but not used:

      CC [M]  .../src/libafs/MODLOAD-2.6.18-419.el5-MP/osi_vfsops.o
    cc1: warnings being treated as errors
    .../src/libafs/MODLOAD-2.6.18-419.el5-MP/osi_vfsops.c: In function 'afs_fill_super':
    .../src/libafs/MODLOAD-2.6.18-419.el5-MP/osi_vfsops.c:167: warning: label 'out' defined but not used

To fix this, put the 'out:' definition under the same conditions as
its users.

rxgk_crypto_rfc3961.c uses 'iterations' in a way where gcc claims it
may be uninitialized:

      CC  .../src/rxgk/rxgk_crypto_rfc3961.lo
    cc1: warnings being treated as errors
    .../src/rxgk/rxgk_crypto_rfc3961.c: In function 'rxgk_derive_tk':
    .../src/rxgk/rxgk_crypto_rfc3961.c:617: warning: 'iterations' may be used uninitialized in this function

This doesn't seem possible to actually use uninitialized, but
initialize it to get rid of the warning.

Change-Id: I68dc4378d03f956c0e81b1dba37ed09d2e396c56
Reviewed-on: https://gerrit.openafs.org/15217
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>