fs: Avoid unnecessary lstat() for -literal
Currently, GetLastComponent() always stats (lstat) the file received as
an argument. However, the status information about this file is only
used when 'literal' is false. That said, skip lstat() when 'literal' is
true.

While here, zero out 'statbuff' at the beginning of GetLastComponent()
just to make sure it won't be used uninitialized in the future.

Change-Id: I3f332c53ae688d6f74276bcdbf2f9f61605ec9ad
Reviewed-on: https://gerrit.openafs.org/14840
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>