afsio: Index into dirName properly in BreakUpPath
In BreakUpPath, dirName is not an array of strings, but an output
argument that is a string (or technically a string array of length 1).
*dirName[dirNameLen] here is effectively dirName[dirNameLen][0], which
is wrong, and usually causes a segfault or other memory weirdness. We
want dirName[0][dirNameLen] to assign to the 'dirNameLen'th character
in the string, so do that instead. But call it (*dirName)[dirNameLen]
to match the other code in this function.

This bug caused segmentation faults when using the 'afsio write'
command for files that do not yet exist. In such cases, the operation
creates a new file, which depends on BreakUpPath() to correctly extract
the parent directory name. That name is needed to obtain the parent’s
FID, a required argument for the RXAFS_CreateFile() RPC.

Other options such as fidwrite, append, and fidappend are unaffected,
since they operate only on existing files. Other callers of
BreakUpPath() within afsio are also safe because they only request the
base name.

This regression was introduced by commit
b8648dbefb3968329d20cad8976ce15947428678.

Change-Id: I108360ec0235dac1acb6c658266af2696f699ab7
Reviewed-on: https://gerrit.openafs.org/15319
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>