Cast LWP event functions to void pointer
When building with Solaris Studio, we receive warnings about the type of
the event passed to the LWP functions.

  CC  /export/home/vagrant/openafs/src/vol/volume.o
  "volume.c", line 662: warning: argument #1 is incompatible with prototype:
     prototype: pointer to void : ".../include/lwp.h", line 273
     argument : pointer to function(enum  {volumeSalvager(7), ...)
                  returning int (E_ARG_INCOMPATIBLE_WITH_ARG_L)

In particular, this happens in src/vol/volume.c because our "events" are
function pointers and the LWP functions expect a pointer to a void.

As needed, cast the event pointer to the expected (void *) to eliminate
the warnings.

[mmeffie: Rebase and update commit message.]

Change-Id: I619a52326914aeff66263e62490de3b6d3b0c9ac
Reviewed-on: https://gerrit.openafs.org/15356
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: Michael Meffie <mmeffie@sinenomine.net>