doc: Add generate-html --stylesheet option
Add the --stylesheet command line option to the generate-html script to
specify the path of the stylesheet to be copied into the generated html
directory. The default value is 'style.css' to retain compatibility.

The --stylesheet option will allow generate-html to work in objdir
builds, and may be used to provide an alternative stylesheet if desired.

Change-Id: I7160d43eb3984e251ba0197bbdde7b16ee893989
Reviewed-on: https://gerrit.openafs.org/16327
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>