fs: Improve 'flush' errno handling
If an error is encountered, 'fs flush' and 'fs flushvolume' try to
print an error message using perror(), but do so after calling
fprintf(), which may change the value of errno.

Also, 'fs flushall' doesn't try to print a reason for an error at all;
it just shows that some kind of error happened.

Change all of these to print an errno-derived error message using
strerror().

Change-Id: I3d7983c6d48068d3b4cb11ae2e6a1d890f8b8ba7
Reviewed-on: https://gerrit.openafs.org/15392
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: Michael Meffie <mmeffie@sinenomine.net>