tests: Add COMMON_LIBS macro to src/tests/Makefile
In order to simplify a future change, introduce a Makefile macro that
defines the common libs used.

Add COMMON_LIBS which will initially be set to just $(LIBS).
Replace $(LIBS) with $(COMMON_LIBS) through the rest of the file.  Add
$(COMMON_LIBS) to the afscp target.

In addition, split the long lines containing the list of libraries.

There are no functional changes in this commit.

Change-Id: Iaf1ef92fdc7fa6ec4adc0582f714416255c1c972
Reviewed-on: https://gerrit.openafs.org/15396
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>