RedHat: Support building RPMs with alternate CellServDB
By default, the makesrpm.pl script downloads a CellServDB file from
grand.central.org. The URL of the file to be downloaded is specified in
the openafs.spec.in file from the source distribution archive.  Network
access can be avoided by specifying a local file to be packaged with the
'cellservdb' positional argument. Due to implementation limitations of
the makesrpm.pl script, the specified file must be in current working
directory and the filename must the same as the CellServDB URL basename
in the openafs.spec.in file (currently 'CellServDB.2022-05-09').

This change adds support for building packages with an alternative
CellServDB, and adds support for specifying a local CellServDB file by
path.

Add the --cellservdb-url option to override the URL used to download the
CellServDB file when the local file is not specified.

Update the sed commands used to generate the openafs.spec file to
update the CellServDB source directive (Source20) when an alternate
URL is given with the --cellservdb-url.

Replace the shell pipe line to extract the CellServDB URL from the
openafs.spec.in file with perl code.  Search for the 'Source20' source
directive instead of searching for the string 'dl/cellservdb' to find
the CellServDB source URL. (After this commit, the source directive
number of the CellServDB source should not be changed without also
changing the number in makesrpm.pl).

Change the handling of the 'cellservdb' positional argument so the local
file is copied to required name in the temporary SOURCES directory. This
also allows us to specify the local file by path, instead of requiring
the local CellServDB file to be in the current working directory when
running makesrpm.pl.

The --cellservdb-url option and the 'cellservdb' argument are not
exclusive. If both are specified, the URL will be set in the generated
openafs.spec file to indicate the upstream location of the packaged
CellServDB file provided by the 'cellservdb' positional argument.

Change-Id: If8818bcdee8500a164f992f42df3c4f9a587d0cd
Reviewed-on: https://gerrit.openafs.org/15405
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>